+++ CHANGES.md
@@ -4,6 +4,11 @@
 
 Released: --
 
+- Remove the deprecated standalone decorators: `input`, `output`, `doc`, and `auth_required`.
+  Use app/blueprint decorators instead (e.g. `input()` -> `app.input()`/`bp.input()`).
+- Deprecate the following parameters ([issue #237][issue_237]):
+  - `role` in `app.auth_required()`/`bp.auth_required()`, use `roles` and always pass a list.
+  - `tag` in `app.doc()`/`bp.doc()`, use `tags` and always pass a list.
 - Add a base class (`apiflask.scaffold.APIScaffold`) for common logic of `APIFlask` and
   `APIBlueprint`, move route decorators and API-related decorators to this base class
   to improve the IDE auto-completion ([issue #231][issue_231]).
@@ -15,6 +20,7 @@ Released: --
 [issue_211]: https://github.com/greyli/apiflask/issues/211
 [issue_231]: https://github.com/greyli/apiflask/issues/231
 [pull_232]: https://github.com/greyli/apiflask/pull/232
+[issue_237]: https://github.com/greyli/apiflask/issues/237
 
 
 ## Version 0.12.0